+++ BioFVM/BioFVM_basic_agent.h
@@ -85,8 +85,8 @@ class Basic_Agent
 	void update_voxel_index();
 	
 	std::vector<double> * internalized_substrates; 
+	std::vector<double> * released_fractions_at_death; 
 	bool use_internal_densities_as_targets; 
-	bool release_internalized_substrates_at_death; 
 	
 	void set_internal_uptake_constants( double dt ); // any time you update the cell volume or rates, should call this function.