+++ core/PhysiCell_phenotype.h
@@ -75,10 +75,11 @@
 
 #include "../BioFVM/BioFVM.h" 
 
+#include "../modules/PhysiCell_settings.h"
+
 using namespace BioFVM; 
 
 namespace PhysiCell{
-
 class Cell;
 class Cycle_Model; 
 class Phenotype; 
@@ -448,11 +449,28 @@ class Cell_Functions
 	void (*contact_function)(Cell* pMyself, Phenotype& my_phenotype, 
 		Cell* pOther, Phenotype& other_phenotype, double dt ); 
 		
-	void (*insternal_substrate_function)(Cell* pCell, Phenotype& phenotype , double dt ); 
+	void (*internal_substrate_function)(Cell* pCell, Phenotype& phenotype , double dt ); 
 	void (*molecular_model_function)(Cell* pCell, Phenotype& phenotype , double dt ); 
-		
 	
 	Cell_Functions(); // done 
+};
+
+class Bools
+{
+	public:
+		std::vector<bool> values; 
+		std::unordered_map<std::string,int> name_map; 
+		std::string& name( int i ); 
+		std::vector<std::string> units; 
+		
+		int size( void ); 
+		void resize( int n ); 
+		int add( std::string name , std::string units , bool value ); 
+		
+		bool& operator[]( int i ); 
+		bool& operator[]( std::string name ); 
+		
+		Bools(); 
 };
 
 class Molecular
@@ -463,23 +481,43 @@ class Molecular
 	
 		// model much of this from Secretion 
 		Molecular(); 
-	
+ 	
 		std::vector<double> internalized_substrates; // we'll set this to replace BioFVM's version  
-
 		std::vector<double> internalized_substrate_release_fractions; 
+		
 		std::vector<double> substrate_creation_rates; 
 		std::vector<double> substrate_use_rates; 
-		// std::vector<bool> 
+		
+		// Boolean, Integer, and Double parameters
+		std::vector<bool> bools; 
+		std::unordered_map<std::string,int> bool_name_map; 
+		std::string& bool_name( int i ); 
+		std::vector<std::string> bool_units; 
+		void resize_bools( int n ); 
+		int add_bool( std::string name , std::string units , bool value ); 
+		bool& access_bool( std::string name ); 
+		
+		std::vector<int> ints; 
+		std::unordered_map<std::string,int> int_name_map; 
+		std::string& int_name( int i ); 
+		std::vector<std::string> int_units; 
+		int& access_int( std::string name ); 
+		
+		std::vector<int> doubles; 
+		std::unordered_map<std::string,int> double_name_map; 
+		std::string& double_name( int i ); 
+		std::vector<std::string> double_units; 
+		double& access_double( std::string name ); 
 		
 		// use this to properly size the secretion parameters to the microenvironment in 
 		// pMicroenvironment
 		void sync_to_current_microenvironment( void ); // done 
 		
 		void advance( Basic_Agent* pCell, Phenotype& phenotype , double dt ); 
+			// has to go somewhere else 
 		
 		// use this to properly size the secretion parameters to the microenvironment 
 		void sync_to_microenvironment( Microenvironment* pNew_Microenvironment ); // done 
-		
 		
 };
 
@@ -498,6 +536,8 @@ class Phenotype
 	Motility motility; 
 	Secretion secretion; 
 	
+	Molecular molecular; 
+	
 	Phenotype(); // done 
 	
 	void sync_to_functions( Cell_Functions& functions ); // done 
@@ -505,40 +545,6 @@ class Phenotype
 	// make sure cycle, death, etc. are synced to the defaults. 
 	void sync_to_default_functions( void ); // done 
 };
-
-/*
-class Microenvironment_Options
-{
- private:
- 
- public: 
-	Microenvironment* pMicroenvironment;
-	std::string name; 
- 
-	std::string time_units; 
-	std::string spatial_units; 
-	double dx;
-	double dy; 
-	double dz; 
-	
-	bool outer_Dirichlet_conditions; 
-	std::vector<double> Dirichlet_condition_vector; 
-	
-	bool simulate_2D; 
-	std::vector<double> X_range; 
-	std::vector<double> Y_range; 
-	std::vector<double> Z_range; 
-	
-	Microenvironment_Options(); 
-	
-	bool calculate_gradients; 
-};
-
-extern Microenvironment_Options default_microenvironment_options; 
-extern Microenvironment microenvironment;
-
-void initialize_microenvironment( void ); 
-*/
 
 };